#188 refactor: dto 설명 수정
